Added soundfile play Ugen (strplay) and a new thread running over shared memory bridge to prefech blocks of audio. Things to fix: original apps/test demo needs to be activated with a button on the GUI, maybe build the GUI *after* audio is ready, Cmd-Q causes crash -- needs a handler to shut down Arco before quitting, it would be cool to "short circuit" communications between audio thread and fileio threads by pushing messages directly onto shared memory queues rather than sending them to the main O2 process.

+++ arco/src/ugen.cpp
@@ -29,7 +29,7 @@ int set_control_service(const char *ctrlservice)
         control_service_addr_len = 0;
         return 1;  // error return
     }
-    control_service_addr_len = csalen + 2;
+    control_service_addr_len = (int) (csalen + 2);
     control_service_addr[0] = '!';
     strcpy(control_service_addr + 1, ctrlservice);
     strcat(control_service_addr, "/");
